api/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3

# Face recognition and image processing
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

